Se actualiza endpointsy otros
