+++ Backend/requirements.txt
@@ -10,3 +10,4 @@ SQLAlchemy==1.4.42
 marshmallow-sqlalchemy==0.28.1
 redis==4.3.4
 Werkzeug==2.0.2
+psycopg2-binary==2.9.4
